// src/app/shared/services/application-css-var-names.enum.ts
export enum ApplicationCssVarNames {
    APPLICATION_BACKGROUND = '--application-background',

    THEME_PRIMARY = '--ngx-fluent-design-theme-primary',
    THEME_SECONDARY = '--ngx-fluent-design-secondary',
    THEME_TERTIARY = '--ngx-fluent-design-theme-tertiary',
    THEME_LIGHT = '--ngx-fluent-design-theme-light',
    THEME_LIGHTER = '--ngx-fluent-design-theme-lighter',
    THEME_LIGHTER_ALT = '--ngx-fluent-design-theme-lighter-alt',
    THEME_DARK = '--ngx-fluent-design-theme-dark',
    THEME_DARKER = '--ngx-fluent-design-theme-darker',
    THEME_DARK_ALT = '--ngx-fluent-design-theme-dark-alt',

    THEME_FOREGROUND_BLACK = '--ngx-fluent-design-foreground-black',
    THEME_FOREGROUND_NEUTRAL_DARK = '--ngx-fluent-design-foreground-neutral-dark',
    THEME_FOREGROUND_NEUTRAL_PRIMARY = '--ngx-fluent-design-foreground-neutral-primary',
    THEME_FOREGROUND_NEUTRAL_PRIMARY_ALT = '--ngx-fluent-design-foreground-neutral-primary-alt',
    THEME_FOREGROUND_NEUTRAL_SECONDARY = '--ngx-fluent-design-foreground-neutral-secondary',
    THEME_FOREGROUND_NEUTRAL_TERTIARY = '--ngx-fluent-design-foreground-neutral-tertiary',
    THEME_FOREGROUND_WHITE = '--ngx-fluent-design-foreground-white',

    THEME_BACKGROUND_NEUTRAL_TERTIARY_ALT = '--ngx-fluent-design-background-neutral-tertiary-alt',
    THEME_BACKGROUND_NEUTRAL_DARK = '--ngx-fluent-design-background-neutral-dark',
    THEME_BACKGROUND_NEUTRAL_QUATERNARY_ALT = '--ngx-fluent-design-background-neutral-quaternary-alt',
    THEME_BACKGROUND_NEUTRAL_LIGHT = '--ngx-fluent-design-background-neutral-light',
    THEME_BACKGROUND_NEUTRAL_LIGHTER = '--ngx-fluent-design-background-neutral-lighter',
    THEME_BACKGROUND_NEUTRAL_LIGHT_ALT = '--ngx-fluent-design-background-neutral-light-alt',
}